requirements.txt
aiogram==2.22.2
Telethon==-1.25.4
coc.py==2.1.3
